Remove `--cachedir` flag
This is only used for caching the `fish_test_helper` binary.
Having changed when this is built, it is no longer necessary to have a cachedir
to cache the binary between the individual tests.

The cachedir would still have value between invocations of `test_driver.py`, but
that has the issue of possibly not rebuilding if the source file changes.
Given that building takes under 100 ms on modern systems, we consider it
unnecessary to keep support for it.