xtask: add `html-docs` task
This task is a bit annoying to implement because `sphinx-build` depends
on `fish_indent`, so that needs to be built, and the binary location
added to `PATH`.

Building `fish_indent` can be avoided by setting the
`FISH_INDENT_FOR_BUILDING_DOCS` environment variable to the path of an
existing `fish_indent`.

The xtask requires specifying the path to the directory where the HTML
files should be put.

Use the new xtask in CMake builds.