Switch to builtin gettext implementation
This completely removes our runtime dependency on gettext. As a
replacement, we have our own code for runtime localization in
`src/wutil/gettext.rs`. It considers the relevant locale variables to
decide which message catalogs to take localizations from. The use of
locale variables is mostly the same as in gettext, with the notable
exception that our code will try to get a different variant of a
language before giving up on localizing into the requested language,
so someone with `LANG=pt_PT.UTF-8` will get messages from the `pt_BR`
catalog, whereas gettext would just give up and return the English
versions of the messages from the source code.

The sourcing of catalogs is completely reworked. Instead of looking for
MO files at runtime, we create catalogs as Rust maps at build time, by
converting PO files into MO data, which is not stored, but immediately
parsed to extract the mappings. From the mappings, we create Rust source
code as a build artifact, which is then macro-included in the crate's
library, i.e. `crates/gettext-maps/src/lib.rs`. The code in
`src/wutil/gettext.rs` includes the message catalogs from this library,
resulting in the message catalogs being built into the executable.

Because the message catalogs take up several megabytes in total, it
might not always be desirable to include them into builds. To make this
configurable, we add cargo features. `localize-all-languages` is enabled
by default, which results in bundling all localizations we have.
When default features are disabled, individual languages can be enabled
via their own feature flags, e.g. `localize-de_DE`.

As a performance optimization, for each language we cache a separate
Rust source file containing its catalog as a map. This allows us to
reuse parsing results if the corresponding PO files have not changed
since we cached the parsing result.

Note that this approach does not eliminate our build-time dependency on
gettext. The process for generating PO files (which uses `msguniq` and
`msgmerge`) is unchanged, and we still need `msgfmt` to translate from
PO to MO. We could parse PO files directly, but these are significantly
more complex to parse, so we use `msgfmt` to do it for us and parse the
resulting MO data.

Advantages of the new approach:
- We have no runtime dependency on gettext anymore.
- The implementation has the same behavior everywhere.
- Our implementation is significantly simpler than GNU gettext.
- We can have localization in cargo-only builds by embedding
  localizations into the code.
  Previously, localization in such builds only worked as long as the
  binary was not moved from the build directory.
- We no longer have to take care of building and installing MO files in
  build systems; everything we need for localization to work happens
  automatically when building fish.
- Reduced overhead when disabling localization, both in compilation time
  and binary size. At the moment, a release build with all features
  disabled results in a binary size of 4.6 MB, whereas when enabling
  only the `localize-all-languages` feature, the resulting size is 9.5
  MB.

Disadvantages of this approach:
- Our own runtime implementation of gettext needs to be maintained.
- The implementation has a more limited feature set (but I don't think
  it lacks any features which have been in use by fish).
- When the set of languages for which we have localizations changes, we
  need to also update every place where the corresponding per-language
  feature needs to be updated. This was an issue in the previous
  CMake-based approach as well, so it is not much of a change.