Parse PO files directly
Skip the step of first converting PO files to MO files using `msgfmt`,
and instead use a custom parser for the PO files which directly extracts
a map from msgid to msgstr. The parser is quite limited, but it can
parse our current set of PO files.

This approach is less robust than parsing MO files due to the PO format
being significantly more complex to parse. It is also a bit slower.

The main advantage is that we no longer need any gettext tooling
installed on the system for compiling fish with localization support. We
still need such tooling for testing and updating PO files from sources,
so it is mainly useful for use cases which only involve compiling fish
on a system.